refactor: worker posv helper for viction
